requirements.txt
streamlit>=1.36.0
numpy>=1.26.0
pandas>=2.2.0
pillow>=10.4.0
matplotlib>=3.8.0
graphviz>=0.20.3
opencv-python>=4.10.0